If already deleted... just do nothing.

+++ src/net/sourceforge/fenixedu/applicationTier/Servico/teacher/services/DeleteInstitutionWorkTimeByOID.java
@@ -17,7 +17,9 @@ public class DeleteInstitutionWorkTimeByOID extends Service {
     public void run(Integer institutionWorkTimeID) throws ExcepcaoPersistencia {
         InstitutionWorkTime institutionWorkTime = (InstitutionWorkTime) persistentObject.readByOID(InstitutionWorkTime.class,
                         institutionWorkTimeID);
-        
-        institutionWorkTime.delete();
+        if (institutionWorkTime != null) {
+            institutionWorkTime.delete();
+        }
     }
+
 }
